+++ NEWS.md
@@ -1,3 +1,17 @@
+# remap 0.3.2
+
+## Patches
+
+* Reformatted spacing in code (no functional change).
+
+* Patched parallel computing functionality.
+
+* Update ggplot geom_line size parameter to linewidth.
+
+* Fixed gcc-UBSAN errors when finding distances for 0 points.
+
+* Restructured package data to work with sf when sf isn't in search path.
+
 # remap 0.3.1
 
 ## Minor changes